Fixed issue #10 - System.OverflowException: Arithmetic operation resulted in an overflow (64-bit systems) by replacing most calls to ToInt32 with ToInt64 Fixed issue #16 - NuGet error "Package restore is disabled by default." when building by updating to NuGet 2.8.1 Including Vestris.ResourceLib.pdb and Vestris.ResourceLib.xml in release zip for documentation and easier debugging support Updated documentation references in sln

+++ Source/ResourceLib/Accelerator.cs
@@ -22,7 +22,7 @@ namespace Vestris.ResourceLib
             _accel = (User32.ACCEL) Marshal.PtrToStructure(
                 lpRes, typeof(User32.ACCEL));
 
-            return new IntPtr(lpRes.ToInt32() + Marshal.SizeOf(_accel));
+            return new IntPtr(lpRes.ToInt64() + Marshal.SizeOf(_accel));
         }
 
         /// <summary>
