+++ Source/ResourceLib/MenuTemplateItemPopup.cs
@@ -46,7 +46,7 @@ namespace Vestris.ResourceLib
             _header = (User32.MENUITEMTEMPLATE) Marshal.PtrToStructure(
                 lpRes, typeof(User32.MENUITEMTEMPLATE));
 
-            lpRes = new IntPtr(lpRes.ToInt32() + Marshal.SizeOf(_header));
+            lpRes = new IntPtr(lpRes.ToInt64() + Marshal.SizeOf(_header));
             lpRes = base.Read(lpRes);
 
             return _subMenuItems.Read(lpRes);
